--- src/Evaluation/WorkloadAccuracyMeter/index.ts
@@ -1,1 +1,0 @@
-export {WorkloadAccuracyMeter} from "./WorkloadAccuracyMeter"
